rx: Print free and allocated counters as unsigned values
Commit b967654434 (rxdebug-print-values-unsigned-20090512) changed the
output of rxdebug to print counters as unsigned values.  However, that
commit missed changing the format of the "free" and "allocated" packets
counters.

Change the printf format specifiers of the "free" and "allocated" packet
counters to print them as unsigned values.

The free packets counter is stored as an afs_int32, which maybe a long
on some platforms.  This counter was previously cast to an int type in
commit 1d93f2da22eb (rx-warnings-and-prototyping-20010623) to avoid a
format specifier warning.  Update the cast to an unsigned int to be
clear this is an unsigned value.

Reviewed-on: https://gerrit.openafs.org/15608
Reviewed-by: Andrew Deason <adeason@sinenomine.net>
Reviewed-by: Mark Vitale <mvitale@sinenomine.net>
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>
(cherry picked from commit bee056613311c2a95f946c8eb25b6def97e7ec29)

Change-Id: I14376240c059bb833648609e131719ada6bb74a8