doc: Add section 3 man pages to AdminRef
The section 3 man pages are missing in the AdminRef document, so add
section 3 to the Makefile and perl script which generates the AdminRef
doc.

Currently, the only man page in section 3 is the AFS::ukernel, so only
one page is missing.

Reviewed-on: https://gerrit.openafs.org/15828
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>
Tested-by: BuildBot <buildbot@rampaginggeek.com>
(cherry picked from commit 7dc8a7ddc204766f1c083919c8732920b0552a68)

Change-Id: I960a671f1b02789f7427dba15459e9f2568d152b