rx: Remove includes for MIN/MAX in rx_packet.h
The commit:
    "opr: replace MIN/MAX macros with opr_min/opr_max"
       (Change I2d7b54193ec91f7ead9c5c5f714d9a8bc7533bf7)
replaced all uses of the MIN and MAX macros with opr_min and opr_max.

The include for sys/param.h and sys/sysmacros.h in rx_packet.h are no
longer needed for MIN/MAX.

Remove the preprocessor conditionals and the includes for param.h and
sysmacros.h.

Change-Id: I8f287b012e17ec5cfd6a44525ecdb5e525e21be5