WINNT: Replace min/max macros with opr_min/opr_max
The commit:
    "opr: replace MIN/MAX macros with opr_min/opr_max"
       (Change I2d7b54193ec91f7ead9c5c5f714d9a8bc7533bf7)
replaced all uses of the MIN and MAX macros with opr_min and opr_max.

We deferred the changes for the WINNT source.

Since we now have a consolidated version of opr_min and opr_max,
replace the min/max macros with opr_min and opr_max.

Remove the definitions for the min and max macros, except for the files
in src/WINNT (which can be handled in a future commit).

Change-Id: I69e2c641c25e4e401e0d0525371bca8e69e2ca50