Replace min/max macros with opr_min/opr_max
The commit:
    "opr: replace MIN/MAX macros with opr_min/opr_max"
       (Change I2d7b54193ec91f7ead9c5c5f714d9a8bc7533bf7)
replaced all uses of the MIN and MAX macros with opr_min and opr_max.

Since we now have a consolidated version of opr_min and opr_max,
replace the min/max macros with opr_min and opr_max.

Remove the definitions for the min and max macros, except for the files
in src/WINNT (which can be handled in a future commit).

There are a few exceptions where we needed to leave the min/max macros.

    everything in src/external/heimdal/
    src/tests/create-files.c
    src/tests/snprintf.c

Change-Id: Ib7c91b4e59aaf76ab77f293016496d6179c23c88