tests: Fix manpage tests for objdir builds
The manpage tests have a couple of problems when running for objdir
builds:

- We try to specify './tests-lib/perl5' as a directory to find our
  helper library. However, the cwd when we're running the tests is in
  an objdir build, where the helper library is in the srcdir. Fix this
  by using the SOURCE env var specified by the tests wrapper.

- All of these tests specify the directory in which to find the man
  pages in a subdir of BUILD, but our manpages are located in the src
  dir (since they are built by regen.sh, not by configure/make). Fix
  this by specifying a SOURCE-based directory instead.

To avoid needing to make the same change for each of these tests, also
refactor the manpage tests so each test only needs to specify the
subdirectory and command name, and get rid of some of the common
boilerplate.

Reviewed-on: https://gerrit.openafs.org/13940
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Benjamin Kaduk <kaduk@mit.edu>
(cherry picked from commit 0b8b6683fb525bbeaf118014beb2371e0cf23d90)

 Conflicts:
	tests/bozo/bos-man-t
	tests/bucoord/backup-man-t
	tests/kauth/kas-man-t
	tests/ptserver/pts-man-t
	tests/venus/fs-man-t
	tests/volser/vos-man-t

cwills@sinenomine.net
conflicts are due the different order  of commits.  In master the
commit:
    "tests: Fix manpage tests for objdir builds" (0b8b6683fb525)
is applied before
    "tests: Accommodate c-tap-harness 4.7" (624219a1b2192).
In 1.8.x the commit:
    "tests: Accommodate c-tap-harness 4.7" (0453c15b2a2cb)
is already present.

Change-Id: I065c6e4e16619bd90c8cf8d28fa5b5cddd67c410