afs: Properly free afs_nfsclient_getcreds xdr bufs
Currently, afs_nfsclient_getcreds() has an unusual method of trying to
free xdr-allocated memory for the output arguments of PAGCB_GetSysName()
and PAGCB_GetCreds(). Some items of the returned array are freed or
assigned elsewhere as we work through the array, and then we free the
array itself with afs_osi_Free(). We don't use xdr_free(), because that
would free some items we already freed or copied a reference to in
another struct.

This means that we don't free the array if we 'goto done' before calling
afs_osi_Free(), which can happen if PAGCB_GetSysName/PAGCB_GetCreds
return an error but still manage to alloc their output arguments.

To avoid this, use xdr_free in the 'goto done' destructor. But to avoid
freeing items we've already freed or referenced elsewhere, set each xdr
field to NULL when it's freed or copied away.

Reviewed-on: https://gerrit.openafs.org/16252
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Mark Vitale <mvitale@sinenomine.net>
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>
(cherry picked from commit 2fa1835033c12b26f9b036ec0711b5ba1d584d06)

 Conflicts:
	src/afs/afs_nfsclnt.c -   In the master version of this commit
	      (2fa1835033c1) there is a change that removes a call to
	      afs_osi_Free() after a non-zero return from afs_GetUser().
	      This afs_osi_Free() call was introduced in the commit:
	         "cb719a093200 afs: Allow afs_GetUser to return errors"
	      which is not applicable for 1.8.x at this time.

Change-Id: I2675a91617a2587d0600fbb2f30972c76e68c7fd