Linux: dkms set dkms.conf CLEAN parameter to true
The dkms 3.1.8 commits:
    (754f26f93a50) "dkms: omit deprecation warning for CLEAN=true"
    (cdad56500434) "dkms: deprecate CLEAN"
    (f93a40aec3b9) "dkms: stop running the CLEAN command"
made the CLEAN parameter obsolete and cause dkms to issue a deprecation
message (without failing the build) if the value of CLEAN is not the
literal string `true`. Since dkms v0.2.18, dkms performs builds in a
temporary build directory that created and populated prior to the build
step.

The dkms.conf generated by openafs.spec currently includes the
statement:
    CLEAN="if [ -e src/libafs/Makefile ]; then make -C src/libafs clean; else true; fi"

With commit 754f26f93a50, dkms (version 3.1.8 and newer) emits a
deprecation message because the value for this CLEAN setting is not
`true`.

    ....
    Building module(s)...
    Signing module /var/lib/dkms/openafs/1.9.2-1.224.g7f45a8.el9/build/openafs.ko
    Deprecated feature: CLEAN (/var/lib/dkms/openafs/1.9.2-1.224.g7f45a8.el9/source/dkms.conf)
    Installing /lib/modules/5.14.0-70.26.1.el9_0.x86_64/extra/openafs.ko.xz

Researching the history of dkms, the original commit (tag v0.22.04)
would perform the build [for the openafs kernel module] within the
openafs source tree, which would require performing some sort of clean,
with the next commit (tag v0.23.18), dkms was changed to first copy the
source for the openafs into a fresh temporary build area then do the
build withing that directory, so performing a clean before the build is
unnecessary.

Update the RedHat/openafs.spec.in file to create dkms.conf with:
    CLEAN=true

NOTE: Prior to 3.1.8, the value for CLEAN= is a command that is executed
before the build step, so setting the value to true simply executes the
'true' command.

Reviewed-on: https://gerrit.openafs.org/16386
Reviewed-by: Michael Laß <bevan@bi-co.net>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>
Tested-by: BuildBot <buildbot@rampaginggeek.com>
(cherry picked from commit 8849df9f895feac6231d05a07f9c9e8a53f5b0a6)

Change-Id: I9d6d4e9ba4fcd26540005deeb34218be01352e5d