Linux: Refactor afs_linux_write_end()/begin()
Refactor afs_linux_write_end() and afs_linux_write_begin() to separate
the folio and non-folio related code.  This allows reducing some of the
preprocessor statements that are within the body of the functions.

The non-folio related code should be stable, while there are future
changes in the folio related code.

There are no functional changes in this commit.

Change-Id: I1d167f9323165ffe203a022723320a8825e7735b